This Spring Boot Weather App fetches real-time weather data from OpenWeatherMap API using RestTemplate and updates the Thymeleaf UI dynamically. It uses Lombok for POJOs, injects the API key via application.properties, and references two HTML files from GitHub for the frontend.

+++ LibraryMgmt/src/main/java/com/project/LibraryMgmt/entity/Publisher.java
@@ -17,7 +17,7 @@ import java.util.Set;
 public class Publisher {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int publisherId;
+    private long publisherId;
 
     @NotNull
     @Column(unique = true)
